third_party/blink/web_tests/platform/generic/fast/dom/Range/insertNode-empty-fragment-crash-expected.txt

Test Range.insertNode(fragment) when the fragment is empty and the range is collapsed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS r.toString() is ""
PASS p.childNodes.length is 2
PASS p.childNodes[0] is t1
PASS p.childNodes[1] is t2
PASS r.collapsed is true
PASS r.startContainer is p
PASS r.startOffset is 1
PASS r.endContainer is p
PASS r.endOffset is 1
PASS r.toString() is ""
PASS successfullyParsed is true

TEST COMPLETE
